fix(test): put scratch dir under CARGO_TARGET_TMPDIR or CARGO_TARGET_DIR
Since Cargo 1.54, `$CARG0_TARGET_TMPDIR` points at a directory in the
`target` directory for tests and benches to use. Prefer to use this, if
it's available.

If not, use `$CARGO_TARGET_DIR` if it's set to honor a custom target
directory being specified by the user (e.g., `cargo test --target-dir
...`).

Otherwise, fallback to the previous use of target/ as the base
directory.